move env parse
